requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.9
pydantic-settings==2.10.1
email-validator==2.2.0
pandas==2.3.2
python-dotenv==1.1.1
kafka-python==2.1.2
